refactor: reorganize features and unify services
