+++ src/components/SectionsList.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, type FC } from 'react';
 import LoadingScreen from './LoadingScreen';
 import { Play, CheckCircle2, Book } from 'lucide-react';
-import { fetchSections } from '../services/courseService.js';
+import { fetchSections } from '../services/courseService';
 import { useLoadData } from '../hooks/useLoadData';
 import { getSectionProgressPercent } from '../services/progressService'
