+++ src/features/account/MyAccount.tsx
@@ -1,21 +1,20 @@
 import { FC, useState, useRef, useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { User, Shield, LogOut, Check, Trophy } from 'lucide-react'
-import { useAuth } from './SupabaseAuthProvider'
-import { isAdmin } from '../utils/adminUtils.js'
-import { findOrCreateUserProfile } from '../services/authService.js'
-import LoadingScreen from './LoadingScreen'
-import AdminPanelButton from './AdminPanelButton'
-import AccountHeader from './account/AccountHeader'
-import AccountStats from './account/AccountStats'
-import AccountProgress from './account/AccountProgress'
-import SectionProgressList from './account/SectionProgressList'
-import SummaryCards from './account/SummaryCards'
-import useChapterStats from '../hooks/useChapterStats'
-import useUserProgress from '../hooks/useUserProgress'
-import ProgressSummary from './account/ProgressSummary'
-import { getCurrentUser } from '../services/authService.js'
-import { getFullUserProgress } from '../services/progressService.js'
+import { useAuth } from '../../components/SupabaseAuthProvider'
+import { isAdmin } from '../../utils/adminUtils.js'
+import { findOrCreateUserProfile, getCurrentUser } from '../../services/authService'
+import LoadingScreen from '../../components/LoadingScreen'
+import AdminPanelButton from '../admin/AdminPanelButton'
+import AccountHeader from './AccountHeader'
+import AccountStats from './AccountStats'
+import AccountProgress from './AccountProgress'
+import SectionProgressList from './SectionProgressList'
+import SummaryCards from './SummaryCards'
+import useChapterStats from '../../hooks/useChapterStats'
+import useUserProgress from '../../hooks/useUserProgress'
+import ProgressSummary from './ProgressSummary'
+import { getFullUserProgress } from '../../services/progressService'
 
 interface MyAccountProps {
   onBackToHome: () => void
